fem_app/requirements.txt
anastruct>=1.5.0
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
plotly>=5.18.0
pandas>=2.1.0
pyyaml>=6.0.1
requests>=2.31.0
pytest>=7.4.0
playwright>=1.41.0
